+++ metadata.rb
@@ -4,6 +4,7 @@ maintainer_email 'rbradberry@simplereach.com'
 license          'All rights reserved'
 description      'Installs/Configures dse-chef'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.0.97'
+version          '0.0.99'
 
 depends 'java'
+depends 'apache2'