+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG for dse-chef
 
-## 0.1.0:
+## 0.0.99
+* adding server search to opscenter-agent recipe (node[:datastax][:opscenter][:server_role] attribute)
+* agent configuration will warn and exit when server cannot be found or too many are found
 
+## 0.0.98
+* opscenter-server recipe
+
+## 0.1.0:
 * Initial release of dse-chef